+++ resources/views/tailwind/form-input.blade.php
@@ -1,14 +1,18 @@
-<div class="mt-4">
+<div class="@if($type === 'hidden') hidden @else mt-4 @endif">
     <label class="block">
         <x-form-label :label="$label" />
 
         <input {!! $attributes->merge([
             'class' => 'form-input block w-full ' . ($label ? 'mt-1' : '')
         ]) !!}
-            name="{{ $name }}"
-            type="{{ $type }}"
-            value="{{ $value }}"
-        />
+            @if($isWired())
+                wire:model="{{ $name }}"
+            @else
+                name="{{ $name }}"
+                value="{{ $value }}"
+            @endif
+
+            type="{{ $type }}" />
     </label>
 
     @if($hasErrorAndShow($name))