+++ resources/views/tailwind/form-select.blade.php
@@ -2,7 +2,13 @@
     <label class="block">
         <x-form-label :label="$label" />
 
-        <select name="{{ $name }}"
+        <select
+            @if($isWired())
+                wire:model="{{ $name }}"
+            @else
+                name="{{ $name }}"
+            @endif
+
             @if($multiple)
                 multiple
             @endif